Adding validation

+++ data-raw/covidBR.R
@@ -1,33 +1,52 @@
 devtools::load_all(".")
 
-list_files <- list.files(pattern="*.csv", path = "./data-raw")
+# Checks if there is a newer version of dataset
 
-list_files <- stringr::str_c("./data-raw",list_files, sep = "/")
+if(cbr_check4updates()){
 
-unlink(list_files)
+  # Delete files from "inst/extdata" folder
+  list_files <- list.files(pattern="*.csv", path = "./inst/extdata")
+  list_files <- stringr::str_c("./inst/extdata",list_files, sep = "/")
+  unlink(list_files)
 
-resp <- cbr_download(url_file = cbr_download_url(),
-                     dir = "./data-raw",
-                     keep_filename = TRUE)
+  # Request
+  resp <- cbr_download(url_file = cbr_download_url(),
+                       dir = "./inst/extdata",
+                       keep_filename = TRUE)
 
-abs_path_unzip <- stringr::str_c(
-  "./data-raw/",
-  stringr::str_replace(resp$url,
-                       pattern = ".+/(.+)\\.\\w+$",
-                       replacement = "\\1"),
-  ".",
-  tools::file_ext(resp$url)
-)
 
-cbr_unzip(abs_path_unzip,
-          dir = "./data-raw",
-          keep_zip_file = FALSE)
+  # Get absolute path
+  abs_path_unzip <- stringr::str_c(
+    "./inst/extdata/",
+    stringr::str_replace(resp$url,
+                         pattern = ".+/(.+)\\.\\w+$",
+                         replacement = "\\1"),
+    ".",
+    tools::file_ext(resp$url)
+  )
 
-today <- Sys.Date()
 
-date_3_months_ago <- today - months(3)
+  # Unzip
+  cbr_unzip(abs_path_unzip,
+            dir = "./inst/extdata",
+            keep_zip_file = FALSE)
 
-covidBR <- cbr_data(dir = "./data-raw",
-                    date = c(date_3_months_ago, today))
 
-usethis::use_data(covidBR, overwrite = TRUE)
+  # Set the date range
+  today <- Sys.Date()
+  date_3_months_ago <- today - months(3)
+
+
+  # Get database
+  covidBR <- cbr_data(dir = "./inst/extdata",
+                      date = c(date_3_months_ago, today))
+
+
+  # Generate the R dataset
+  usethis::use_data(covidBR, overwrite = TRUE)
+
+}else{
+
+  message("The data is now up to date!")
+
+}
